added views and configured update

+++ Properties/src/main/java/com/example/Properties/Property/Model/Property.java
@@ -30,7 +30,7 @@ public Property() {
         return propertyId;
     }
 
-    public void setPropertyId(Integer propertyId) {
+    public void setPropertyId(long propertyId) {
         this.propertyId = propertyId;
     }
